Move and split the heatpumps module
